Extract EditPersonDescriptor from EditCommand
