+++ src/main/java/seedu/address/logic/parser/EditCommandParser.java
@@ -15,8 +15,8 @@ import java.util.Set;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.model.common.EditPersonDescriptor;
 import seedu.address.model.tag.Tag;
 
 /**
